feat: get result from Task
