+++ README.md
@@ -15,42 +15,30 @@ go get github.com/KenjiHosaka/goctopus
 ## How to use
 ### Simple
 ```golang
-outputs, err := goctopus.Orchestrate(
-	context.Background(), 
-	goctopus.NewTask[bool](func() (bool, error) {
-		return true, nil
-	}).Run(), 
-	goctopus.NewTask[string](func() (string, error) {
-		return "result", nil
-	}).Run(), 
-	goctopus.NewTask[int](func() (int, error) {
-		return 0, nil
-	}).Run(), 
-)()
-```
-
-### Get Result
-```golang
 task1 := goctopus.NewTask[bool](func() (bool, error) {
-	time.Sleep(10 * time.Millisecond)
 	return true, nil
 })
 
-outputs, err := goctopus.Orchestrate(
+task2 := goctopus.NewTask[string](func() (string, error) {
+	return "result", nil
+})
+
+err := goctopus.Orchestrate(
 	context.Background(), 
-	task1.Run(),
+	task1.Run(), 
+	task2.Run(), 
 )()
-res, exist := goctopus.FindResult(outputs, task1)
-// res: true
+
+// Get Result
+res := task2.Result()
+// res: result
+
 ```
 
 ### Set Timeout
 ```golang
-outputs, err := goctopus.Orchestrate(
-	context.Background(), 
-	goctopus.NewTask[bool](func() (bool, error) {
-		return true, nil
-	}).Run(), 
+err := goctopus.Orchestrate(
+	// ...
 )(goctopus.TimeOut{
 	Duration: 1 * time.Second,
 })
