+++ goctopus.go
@@ -2,21 +2,13 @@ package goctopus
 
 import (
 	"context"
-	"math/rand"
 	"time"
 )
 
-type taskResult struct {
-	taskID uint32
-	value  any
-}
-
-type AsyncTask func(ctx context.Context) (taskResult, error)
-
-type OrchestrationFunc func(...Option) (Outputs, error)
+type OrchestrationFunc func(...Option) error
 
 func Orchestrate(ctx context.Context, tasks ...AsyncTask) OrchestrationFunc {
-	return func(opts ...Option) (Outputs, error) {
+	return func(opts ...Option) error {
 		options := Options{}
 		for _, opt := range opts {
 			opt.apply(&options)
@@ -34,7 +26,7 @@ func Orchestrate(ctx context.Context, tasks ...AsyncTask) OrchestrationFunc {
 
 		recoverCh := make(chan interface{}, len(tasks))
 		errCh := make(chan error, len(tasks))
-		doneCh := make(chan taskResult, len(tasks))
+		doneCh := make(chan struct{}, len(tasks))
 
 		for _, task := range tasks {
 			go func(t AsyncTask) {
@@ -45,70 +37,29 @@ func Orchestrate(ctx context.Context, tasks ...AsyncTask) OrchestrationFunc {
 					}
 				}()
 
-				res, err := t(c)
+				err := t(c)
 
 				if err != nil {
 					errCh <- err
 					return
 				}
 
-				doneCh <- res
+				doneCh <- struct{}{}
 			}(task)
 		}
 
-		results := make(map[uint32]Output, len(tasks))
 		for i := 0; i < len(tasks); i++ {
 			select {
 			case <-c.Done():
-				return nil, c.Err()
-			case res := <-doneCh:
-				results[res.taskID] = Output{
-					result: res.value,
-				}
+				return c.Err()
+			case <-doneCh:
 			case err := <-errCh:
-				return nil, err
+				return err
 			case r := <-recoverCh:
 				panic(r)
 			}
 		}
 
-		return results, nil
-	}
-}
-
-func NewTask[T any](f func() (T, error)) Task[T] {
-	return Task[T]{
-		id: rand.Uint32(),
-		f:  f,
-	}
-}
-
-type Task[T any] struct {
-	id uint32
-	f  func() (T, error)
-}
-
-func (t Task[T]) Run() AsyncTask {
-	return func(ctx context.Context) (taskResult, error) {
-		errCh := make(chan error, 1)
-		resCh := make(chan T, 1)
-		go func() {
-			res, e := t.f()
-			if e != nil {
-				errCh <- e
-			}
-
-			resCh <- res
-		}()
-
-		select {
-		case <-ctx.Done():
-			res := <-resCh
-			return taskResult{taskID: t.id, value: res}, ctx.Err()
-		case err := <-errCh:
-			return taskResult{}, err
-		case res := <-resCh:
-			return taskResult{taskID: t.id, value: res}, nil
-		}
+		return nil
 	}
 }
